refactor: extract tmux_session_exists to shared helper
Move duplicate tmux_session_exists() from listen.py to
agents/tmux.py as a module-level function.

Built by dotdev.dev
